backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pgvector
python-dotenv
pydantic
python-multipart
openai
dashscope